+++ src/overture_airflow_provider/_wherobots.py
@@ -18,7 +18,6 @@ try:
 except ImportError:
     WHEROBOTS_AVAILABLE = False
 
-MAX_TIMEOUT_HOURS = 8
 WHEROBOTS_PROVIDER = "com.wherobots.awssdk.auth.WherobotsAssumeRoleCredentialsProvider"
 _API_SUBDOMAIN_PREFIX = "api."
 
@@ -273,6 +272,7 @@ def build_wherobots_operator_kwargs(
     spark_cluster_size: str,
     spark_cluster_desired_worker_cores: str,
     spark_cluster_desired_workers: str,
+    max_timeout_hours: str,
     wherobots_role_arn: str,
     task_id: str,
     resolve_region: bool = True,
@@ -395,7 +395,7 @@ def build_wherobots_operator_kwargs(
         "runtime": runtime_name,
         "poll_logs": poll_logs,
         "polling_interval": polling_interval,
-        "timeout_seconds": (3600 * MAX_TIMEOUT_HOURS),
+        "timeout_seconds": (3600 * int(max_timeout_hours)),
         "region": region_val,
         "environment": environment,
     }
@@ -436,6 +436,7 @@ def execute_wherobots_job(
     spark_cluster_size: str,
     spark_cluster_desired_worker_cores: str,
     spark_cluster_desired_workers: str,
+    max_timeout_hours: str,
     wherobots_role_arn: str,
     task_id: str,
     context,
@@ -458,6 +459,7 @@ def execute_wherobots_job(
         spark_cluster_size=spark_cluster_size,
         spark_cluster_desired_worker_cores=spark_cluster_desired_worker_cores,
         spark_cluster_desired_workers=spark_cluster_desired_workers,
+        max_timeout_hours=max_timeout_hours,
         wherobots_role_arn=wherobots_role_arn,
         task_id=task_id,
         resolve_region=True,